Move back common/reducers to its original path.
