+++ src/store/reducers.ts
@@ -3,11 +3,12 @@ import { History } from 'history';
 import { combineReducers } from 'redux';
 import { ActionType } from 'typesafe-actions';
 
-import { StoreState } from '../../util/types';
-import * as actions from '../actions';
-import { blockchain } from '../blockchain/reducers';
-import { relayer } from '../relayer/reducers';
-import { ui } from '../ui/reducers';
+import { StoreState } from '../util/types';
+
+import * as actions from './actions';
+import { blockchain } from './blockchain/reducers';
+import { relayer } from './relayer/reducers';
+import { ui } from './ui/reducers';
 
 export type RootAction = ActionType<typeof actions>;
